+++ lib/src/code_generators/swagger_additions_generator.dart
@@ -124,6 +124,11 @@ String? _dateToJson(DateTime? date) {
 
   return '\$year-\$month-\$day';
   }
+
+  class Wrapped<T> {
+  final T value;
+  const Wrapped.value(this.value);
+}
 ''';
   }